flask/requirements.txt
black==22.8.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
python-dotenv==0.21.0
SQLAlchemy==1.4.41
tomli==2.0.1
Werkzeug==2.2.2
uWSGI==2.0.20

